Initial stab at refactoring Ill and Aeon links.  Broke some things but getting there
